Fix compilation problem on linux64 (needed intptr_t cast)
git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4795 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ compiler/ifa/sym.cpp
@@ -417,7 +417,7 @@ struct AlignOf##_t {      \
   _t b;                  \
 }
 
-#define ALIGNOF(_t) ((int)&(((struct AlignOf##_t *)0)->b))
+#define ALIGNOF(_t) ((int)(intptr_t)&(((struct AlignOf##_t *)0)->b))
 
 MAKE_ALIGNOF(bool);
 MAKE_ALIGNOF(uint8);
